[Reorg] Split main() into own file
