+++ src/main.cpp
@@ -1,0 +1,256 @@
+/***** BEGIN LICENSE BLOCK *****
+
+BSD License
+
+Copyright (c) 2005-2015, NIF File Format Library and Tools
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions
+are met:
+1. Redistributions of source code must retain the above copyright
+   notice, this list of conditions and the following disclaimer.
+2. Redistributions in binary form must reproduce the above copyright
+   notice, this list of conditions and the following disclaimer in the
+   documentation and/or other materials provided with the distribution.
+3. The name of the NIF File Format Library and Tools project may not be
+   used to endorse or promote products derived from this software
+   without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE AUTHOR ``AS IS'' AND ANY EXPRESS OR
+IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES
+OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED.
+IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY DIRECT, INDIRECT,
+INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT
+NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
+THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+***** END LICENCE BLOCK *****/
+
+#include <QApplication>
+#include <QDir>
+#include <QStack>
+#include <QSettings>
+#include <QLocale>
+#include <QCommandLineParser>
+
+#include <QDesktopServices>
+#include <QUdpSocket>
+#include <QUrl>
+
+#include "nifskope.h"
+#include "nifvalue.h"
+#include "nifmodel.h"
+#include "kfmmodel.h"
+#include "version.h"
+
+
+QCoreApplication * createApplication( int &argc, char *argv[] )
+{
+	// Iterate over args
+	for ( int i = 1; i < argc; ++i ) {
+		// -no-gui: start as core app without all the GUI overhead
+		if ( !qstrcmp( argv[i], "-no-gui" ) ) {
+			return new QCoreApplication( argc, argv );
+		}
+	}
+	return new QApplication( argc, argv );
+}
+
+
+/*
+ *  main
+ */
+
+//! The main program
+int main( int argc, char * argv[] )
+{
+	QScopedPointer<QCoreApplication> app( createApplication( argc, argv ) );
+
+	if ( auto a = qobject_cast<QApplication *>(app.data()) ) {
+
+		a->setOrganizationName( "NifTools" );
+		a->setOrganizationDomain( "niftools.org" );
+		a->setApplicationName( "NifSkope " + NifSkopeVersion::rawToMajMin( NIFSKOPE_VERSION ) );
+		a->setApplicationVersion( NIFSKOPE_VERSION );
+		a->setApplicationDisplayName( "NifSkope " + NifSkopeVersion::rawToDisplay( NIFSKOPE_VERSION, true ) );
+
+		// Must set current directory or this causes issues with several features
+		QDir::setCurrent( qApp->applicationDirPath() );
+
+		// Register message handler
+		//qRegisterMetaType<Message>( "Message" );
+		qInstallMessageHandler( NifSkope::MessageOutput );
+
+		// Register types
+		qRegisterMetaType<NifValue>( "NifValue" );
+		QMetaType::registerComparators<NifValue>();
+
+		// Find stylesheet
+		QDir qssDir( QApplication::applicationDirPath() );
+		QStringList qssList( QStringList()
+			<< "style.qss"
+#ifdef Q_OS_LINUX
+			<< "/usr/share/nifskope/style.qss"
+#endif
+		);
+		QString qssName;
+		for ( const QString& str : qssList ) {
+			if ( qssDir.exists( str ) ) {
+				qssName = qssDir.filePath( str );
+				break;
+			}
+		}
+
+		// Load stylesheet
+		if ( !qssName.isEmpty() ) {
+			QFile style( qssName );
+
+			if ( style.open( QFile::ReadOnly ) ) {
+				a->setStyleSheet( style.readAll() );
+				style.close();
+			}
+		}
+
+		// Set locale
+		QSettings cfg( QString( "%1/nifskope.ini" ).arg( QCoreApplication::applicationDirPath() ), QSettings::IniFormat );
+		cfg.beginGroup( "Settings" );
+		NifSkope::SetAppLocale( cfg.value( "Locale", "en" ).toLocale() );
+		cfg.endGroup();
+
+		// Load XML files
+		NifModel::loadXML();
+		KfmModel::loadXML();
+
+		int port = NIFSKOPE_IPC_PORT;
+
+		QStack<QString> fnames;
+
+		// Command Line setup
+		QCommandLineParser parser;
+		parser.addHelpOption();
+		parser.addVersionOption();
+
+		// Add port option
+		QCommandLineOption portOption( {"p", "port"}, "Port NifSkope listens on", "port" );
+		parser.addOption( portOption );
+
+		// Process options
+		parser.process( *a );
+
+		// Override port value
+		if ( parser.isSet( portOption ) )
+			port = parser.value( portOption ).toInt();
+
+		// Files were passed to NifSkope
+		for ( const QString & arg : parser.positionalArguments() ) {
+			QString fname = QDir::current().filePath( arg );
+
+			if ( QFileInfo( fname ).exists() ) {
+				fnames.push( fname );
+			}
+		}
+
+		// No files were passed to NifSkope, push empty string
+		if ( fnames.isEmpty() ) {
+			fnames.push( QString() );
+		}
+
+		if ( IPCsocket * ipc = IPCsocket::create( port ) ) {
+			//qDebug() << "IPCSocket exec";
+			ipc->execCommand( QString( "NifSkope::open %1" ).arg( fnames.pop() ) );
+
+			while ( !fnames.isEmpty() ) {
+				IPCsocket::sendCommand( QString( "NifSkope::open %1" ).arg( fnames.pop() ), port );
+			}
+
+			return a->exec();
+		} else {
+			//qDebug() << "IPCSocket send";
+			while ( !fnames.isEmpty() ) {
+				IPCsocket::sendCommand( QString( "NifSkope::open %1" ).arg( fnames.pop() ), port );
+			}
+			return 0;
+		}
+	} else {
+		// Future command line batch tools here
+	}
+
+	return 0;
+}
+
+
+
+/*
+*  IPC socket
+*/
+
+IPCsocket * IPCsocket::create( int port )
+{
+	QUdpSocket * udp = new QUdpSocket();
+
+	if ( udp->bind( QHostAddress( QHostAddress::LocalHost ), port, QUdpSocket::DontShareAddress ) ) {
+		IPCsocket * ipc = new IPCsocket( udp );
+		QDesktopServices::setUrlHandler( "nif", ipc, "openNif" );
+		return ipc;
+	}
+
+	return nullptr;
+}
+
+void IPCsocket::sendCommand( const QString & cmd, int port )
+{
+	QUdpSocket udp;
+	udp.writeDatagram( (const char *)cmd.data(), cmd.length() * sizeof( QChar ), QHostAddress( QHostAddress::LocalHost ), port );
+}
+
+IPCsocket::IPCsocket( QUdpSocket * s ) : QObject(), socket( s )
+{
+	QObject::connect( socket, &QUdpSocket::readyRead, this, &IPCsocket::processDatagram );
+}
+
+IPCsocket::~IPCsocket()
+{
+	delete socket;
+}
+
+void IPCsocket::processDatagram()
+{
+	while ( socket->hasPendingDatagrams() ) {
+		QByteArray data;
+		data.resize( socket->pendingDatagramSize() );
+		QHostAddress host;
+		quint16 port = 0;
+
+		socket->readDatagram( data.data(), data.size(), &host, &port );
+
+		if ( host == QHostAddress( QHostAddress::LocalHost ) && (data.size() % sizeof( QChar )) == 0 ) {
+			QString cmd;
+			cmd.setUnicode( (QChar *)data.data(), data.size() / sizeof( QChar ) );
+			execCommand( cmd );
+		}
+	}
+}
+
+void IPCsocket::execCommand( const QString & cmd )
+{
+	if ( cmd.startsWith( "NifSkope::open" ) ) {
+		openNif( cmd.right( cmd.length() - 15 ) );
+	}
+}
+
+void IPCsocket::openNif( const QUrl & url )
+{
+	auto file = url.toString();
+	file.remove( 0, 4 );
+
+	openNif( file );
+}
+
+void IPCsocket::openNif( const QString & url )
+{
+	NifSkope::createWindow( url );
+}
